% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_across_sample_ratios}
\alias{calculate_across_sample_ratios}
\title{Calculate across sample ratios}
\usage{
calculate_across_sample_ratios(
  data,
  metadata = NULL,
  validRatios = NULL,
  abundanceCol = "PG.Intensity"
)
}
\arguments{
\item{data}{output from calculate_SILAC_ratios. In case you did something to this dataframe beforehand (eg filtering) make sure it is not grouped tibble before passing it to this function.}

\item{metadata}{tibble containing columns Run, Sample and replicate. Ratio calculation will than be only performed across samples (replicate wise). If no metadata given, all possible Run combinations are calculated}

\item{validRatios}{output from requantify_sanity_check, only neccessary if requantify was used}

\item{abundanceCol}{Identifier of the abundance column(s). Default is "PG.Intensity" which is the PG intensity output identifier from the calculate_SILAC_ratios function}
}
\value{
tibble with all across sample ratios. If validRatios was given it contains a column "is.valid". is.valid == T, the ratios is valid to use. is.valid == F ratios are returned for experimental reasons but should be removed for further analysis.
}
\description{
Function to calculate across sample ratios
}
\examples{

# example uses bulk data

## for requantify data
data <- clean_DIANN(data) # read in the data
filter_frame <- filter_DIANN(data) # generate filter data frame

silac_ratios <- calculate_SILAC_ratios(data = data, filterSet = filter_frame, useFilter = "requantify", globalReference = "H", 
globalCalculation = "all", PGcalculation = "standard") # calculate SILAC abundances

sane_PGs <- requantify_sanity_check(filterSet = filter_frame) # sanity check requantify combinations (without metadata)
data_across <- calculate_across_sample_ratios(data = silac_ratios, validRatios = sanePGs)


## without requantify: 
data <- clean_DIANN(data) # read in the data
filter_frame <- filter_DIANN(data) # generate filter data frame

silac_ratios <- calculate_SILAC_ratios(data = data, filterSet = filter_frame, useFilter = "standard", globalReference = "H", 
globalCalculation = "all", PGcalculation = "standard") # calculate SILAC abundances
data_across <- calculate_across_sample_ratios(data = silac_ratios)
}
