% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_DIANN}
\alias{clean_DIANN}
\title{Clean DIA-NN output}
\usage{
clean_DIANN(file, contaminants = "CON")
}
\arguments{
\item{file}{Path to report.tsv or report.parquet file}

\item{contaminats}{String defining contaminant protein groups}
}
\value{
DIA-NN output cleaned for contaminants and precursors with charge 1
}
\description{
Removing wrong charges and contaminants from any DIANN output file
}
\examples{
data <- clean_DIANN(filepath)
}
