% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert.20}
\alias{convert.20}
\title{Convert DIA-NN 2.0 output to format in lower versions}
\usage{
convert.20(file)
}
\arguments{
\item{file}{file path}
}
\value{
tibble with Precursor.Id formatted as in earlier DIA-NN versions and removed Channel column.
}
\description{
Converting Precursor.Id and Channel columns from the new output file back to the old format so the functions run. Do not use as stand alone. convert.2.0 wraps around it
}
\examples{
data <- convert.20(path/to/file/report.parquet)
}
