% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_precursor_id}
\alias{convert_precursor_id}
\title{Convert Precursor.Id from DIA-NN 2.0 output}
\usage{
convert_precursor_id(precursor_id, channel)
}
\arguments{
\item{precursor_id}{Precursor.Id}

\item{channel}{Channel}
}
\value{
Converted Precursor.Id
}
\description{
Converting Precursor.Id and Channel columns from the new output file back to the old format so the functions run. Do not use as stand alone. convert_20 wraps around it
}
