% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_DIANN}
\alias{filter_DIANN}
\title{Flagging proteins, precursors and channels that pass filters}
\usage{
filter_DIANN(
  data,
  referenceChannel = "H",
  numberChannels = 2,
  MBR = "off",
  CalcCols = "both",
  ChannelFilt = NA,
  version = "1.8.1"
)
}
\arguments{
\item{data}{Output from clean_DIANN function or raw DIANN output.}

\item{referenceChannel}{String defining which channel was spiked-in to use as a reference. Options are "H", "M" and "L". Default is H.}

\item{numberChannels}{Number of SILAC channels, default is 2}

\item{MBR}{Whether MBR was turned on and off. Default is off. Should only be turned on for Label free data}

\item{CalcCols}{Whether or not Ms1.Translated and Precursor.Translated must both have valid values or not. Default is "both" (non valid values are not left). This is recommended for eg timsTOF pro2 data analysed with DIA-NN 1.8.1. For sc data (and higher DIA-NN Version 1.8.2 beta 22 or higher) set to "any" (at least one of them have to be valid / uses Precursor.Normalised for higher DIA-NN versions). "Precursor.Translated" and "Ms1.Translated" can also be used seperately but is not recommenended}

\item{ChannelFilt}{Q value filter for Channel.Q.Value. If left NA, the default and suggested for DIA-NN 1.8.1 is 0.03. For higher versions the recommendation is ChannelFilt = 0.01}

\item{version}{DIA-NN version used to analyse the raw data. Default and recommended is 1.8.1. Set to 1.9 for higher versions. Recommended DIA-NN settings for 1.8.2 beta 22 or higher are QuantUMS: legacy; turn off MBR; and --channel-spec-norm}
}
\value{
Set of Protein.Groups and precursors per Run that passed filters based on an approach similar to requantify in Maxquant. If the reference passes the filter, other channels are also kept. Proteins identified with Requantify may not be quantified accurately. Therefore, the output also contains informations about how many channels per precursor passed the standard filter and an advanced strict filter that further reduces untrustworthy quantifications. Output contains: Run, Protein.Group, Stripped.Sequence.Charge (Precursor.Id without SILAC Channel information), Reference.Channel (Channel used for Requantify), Channels.Passed.Standard.Filter (Listing all Channels that passed the standard filter, not only the reference one), Channels.Passed.Strict.Filter (Listing all Channels that passed the strict filter, not only the reference one), N.Total.Precursors.Per.Protein.Standard (Number of requantify precursors that passed the standard filter), N.Precursors.With.Complete.SILAC.Channels.Per.Protein.Standard (Number of precursors per proteins where all channels passed the standard filter) and N.Precursors.With.Complete.SILAC.Channels.Per.Protein.Strict (Number of Precursors per Protein where all channels passed the strict filter).Function can also be used for non-SILAC runs. Set reference-Channel = "L" and numberChannels = 1.
}
\description{
Creates a tibble with precursors and channels that pass certain filters. Filters used are the standard DIA-NN filters used in other publications (Global.PG.Q.Value < 0.01) and Channel.Q.Value < 0.03.
}
\examples{
data <- clean_DIANN(filepath)

# for bulk SILAC spike-in data
filter_frame <- filter_DIANN(data)

 # for sc SILAC spike-in data
filter_frame <- filter_DIANN(data, CalcCols = "any")

 # for DIA-NN 1.8.2 beta 22 or higher
 filter_frame <- filter_DIANN(data, CalcCols = "any", ChannelFilt = 0.01, version = "1.9")
}
