% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_pgMatrix}
\alias{generate_pgMatrix}
\title{Generate human readable PG matrix outputs}
\usage{
generate_pgMatrix(PGdata, filterSet)
}
\arguments{
\item{PGdata}{Output from calculate_SILAC_ratios}

\item{filterSet}{Output from filter_DIANN}
}
\value{
Returns Two wide tables: a) pg_matrix: a table with the same format as pg_matrix.tsv output from DIA-NN containing the protein abundances per run and b) peptide_counts: a wide table version of the filter_DIANN output to allow for easily readable quality control. Contains columns Protein.Group, Unique.Peptides (unique stripped peptide sequences without charge state) in all runs and L;M;H peptide counts per run. Differently from the output of filter_DIANN, this dataframe contains peptides, not precursors (no charge state).
}
\description{
Generates wide table versions of protein abundance and quality measurements (number of unique peptides in total as well as per run that do pass the filter). The output of this function is no input for any other function of this pipeline but to enhance easier compatibility with scripts of other users.
}
\examples{
# for bulk
data <- clean_DIANN(data) # read in the data
filter_frame <- filter_DIANN(data) # generate filter data frame

silac_ratios <- calculate_SILAC_ratios(data = data, filterSet = filter_frame, useFilter = "requantify", globalReference = "H", 
globalCalculation = "all", PGcalculation = "standard")

pg_matrix <- generate_pgMatrix(silac_ratios, filter_frame)

}
