% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{requantify_sanity_check}
\alias{requantify_sanity_check}
\title{Sanity Check Requantify across sample ratios}
\usage{
requantify_sanity_check(filterSet, metadata = NULL, precursorPerProtein = 1)
}
\arguments{
\item{filterSet}{output from filter_DIANN}

\item{metadata}{tibble containing columns Run, Sample and replicate. Sanity Check will than be only performed across samples (replicate wise). If no metadata given, all possible Run combinations are tested}

\item{precursorPerProtein}{precursors per protein required, should be the same as used in calculate_SILAC_ratios}
}
\value{
tibble with all possible valid combinations.
}
\description{
Function to find sane requantify across sample ratios, strongly advoced to run before calculating across sample ratios
NOT WORKING WITH TRIPPLE SILAC
}
\examples{
example uses bulk data
data <- clean_DIANN(data) # read in the data
filter_frame <- filter_DIANN(data) # generate filter data frame

silac_ratios <- calculate_SILAC_ratios(data = data, filterSet = filter_frame, useFilter = "requantify", globalReference = "H", 
globalCalculation = "all", PGcalculation = "standard") # calculate SILAC abundances

sane_PGs <- requantify_sanity_check(filterSet = filter_frame) # sanity check requantify combinations (without metadata)
}
